Added 3 new sketches to the repo as wells as project README
JoyBox_Midi_buttons & JoyBox_Midi_buttons_2 are slightly more usable
version of the original sketch that allow you to send MIDI notes using
the Joystick. These are really POCs and not really very fun to use.

JoyBox_Midi_Arpeg is an arpeggiator controlled by the joystick and
button. This sketch is still a work in progress but can be pretty fun
to use.